Separated App into multiple components for cleanliness.
