+++ webapp/src/components/common/loadingSpinner/LoadingSpinner.svelte
@@ -1,0 +1,1 @@
+<span>LOADING</span>
