+++ webapp/src/components/common/requestError/RequestError.svelte
@@ -1,0 +1,7 @@
+<script>
+  export let error = null;
+</script> 
+
+{#if error}
+  <span>Oops {error}</span>
+{/if}
