+++ webapp/src/components/common/sidebar/SidebarItem.svelte
